Move admin reports to seperate module
